# internal/httpcontroller/v1/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"time"
)

// AppendLinkRequest defines model for AppendLinkRequest.
type AppendLinkRequest struct {
	Link *string `json:"link,omitempty"`
}

// ArchLink defines model for ArchLink.
type ArchLink struct {
	Error    *string `json:"error,omitempty"`
	FileName *string `json:"fileName,omitempty"`
	Link     *string `json:"link,omitempty"`
}

// ArchLinks defines model for ArchLinks.
type ArchLinks = []ArchLink

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code    *int    `json:"code,omitempty"`
	Details *string `json:"details,omitempty"`
	Error   *string `json:"error,omitempty"`
}

// Links defines model for Links.
type Links = []string

// SetOfLinks defines model for SetOfLinks.
type SetOfLinks struct {
	Links *Links `json:"links,omitempty"`
}

// Task defines model for Task.
type Task struct {
	Created  *time.Time `json:"created,omitempty"`
	Download *string    `json:"download,omitempty"`
	Error    *string    `json:"error,omitempty"`
	Id       *int       `json:"id,omitempty"`
	Links    *ArchLinks `json:"links,omitempty"`
	Status   *string    `json:"status,omitempty"`
}

// AppendLinkJSONRequestBody defines body for AppendLink for application/json ContentType.
type AppendLinkJSONRequestBody = AppendLinkRequest

// SetOfLinksJSONRequestBody defines body for SetOfLinks for application/json ContentType.
type SetOfLinksJSONRequestBody = SetOfLinks